feat: add model testing for predict function with linear regression
